requirements.txt
# dependencies required for Clyde
# a curated selection of dependencies guaranteeing fast asynchronous running of the bot and API
# it will not bloat your computer up with python packages

# bot library (select one, git dependency for selfbot, discord.py proper for bot, these cannot be combined)
# git+https://github.com/dolfies/discord.py-self
# discord.py

# gpt libraries
python-tgpt
g4f

# async fix
nest-asyncio

# anti-token leak
python-dotenv

# api requests
brotli
httpx
flask[async]
